+++ source/codegen/common_helpers.py
@@ -13,6 +13,9 @@ def is_enum(parameter):
 def is_struct(parameter):
   return parameter["type"].startswith("struct")
 
+def is_string_arg(parameter):
+  return parameter['type'] in ['ViChar[]', 'ViInt8[]', 'ViUInt8[]']
+
 def get_underlying_type_name(parameter_type):
   '''Strip away information from type name like brackets for arrays, leading "struct ", etc. leaving just the underlying type name.'''
   return parameter_type.replace("struct ","").replace('[]', '')
@@ -32,7 +35,7 @@ def is_unsupported_struct(parameter):
   return is_struct(parameter) and is_input_parameter(parameter)
 
 def is_unsupported_scalar_array(parameter):
-  if (not is_array(parameter['type'])) or (is_output_parameter(parameter) and get_underlying_type_name(parameter['type']) == 'ViUInt8'):
+  if (not is_array(parameter['type'])) or (is_output_parameter(parameter) and is_string_arg(parameter)):
     return False
   return is_enum(parameter) or get_underlying_type_name(parameter['type']) == 'ViInt16'
 
@@ -94,6 +97,14 @@ def has_viboolean_array_param(functions):
   for function in functions:
     for parameter in functions[function]["parameters"]:
       if parameter['type'] == 'ViBoolean[]':
+        return True
+  return False
+
+def has_enum_array_string_out_param(functions):
+  '''Returns True if atleast one function has output parameter of type ViChar[], ViInt8[] or ViUInt8[] that uses enum'''
+  for function in functions:
+    for parameter in functions[function]["parameters"]:
+      if is_output_parameter(parameter) and is_string_arg(parameter) and is_enum(parameter):
         return True
   return False