+++ source/codegen/metadata/nifake/CHANGES.md
@@ -23,6 +23,8 @@ The following functions, not originally in nimi-python metadata, were newly adde
 	- This function allows testing of ViSession[] input parameter
 - `'GetAnIviDanceWithATwistArray'` 
   -  This function allows testing of ivi-dance-with-a-twist mechanism
+- `'GetArrayViUInt8WithEnum'`
+  - This function allows testing of ViUInt8[] output parameters that use enum
  
 The following functions were tagged with `'init_method': True,` to ensure their generated service handlers register the new session
 with the session_repository.